Initial commit
It gets up to calculating a histogram per UV pixel for each chunk. The
merging of the histograms has not happened as yet, neither has the
median/std calculation per pixel.

Once those two are done, the final step is to calculate thresholds and
flag on the visibilities based on the thresholds.